Clarify validation-query semantics before the new API shape hardens
The latest review comments were right that the newly introduced query-test
counters are cheaper to widen now than in a follow-up, and that the
non-read-only SQL classifier belongs closer to the validation-query flow.
This also trims the CLI coverage down to what those integration tests
actually need by reusing the mock gRPC harness for render/exit behavior.

Constraint: The query-test result fields were introduced in this branch and should be corrected before they become established
Constraint: Keep source/runtime ownership boundaries intact while addressing review feedback
Rejected: Keep u32 counters | preserves unnecessary cast and overflow fallback logic in a fresh contract
Rejected: Keep file-backed CLI query-test fixtures | duplicates app-level behavior coverage with heavier setup
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep source_test_errors_when_required_secret_is_missing file-backed because it validates local manifest/config secret resolution before the gRPC boundary
Tested: make rust-checks
Not-tested: Compatibility against any external out-of-tree client generated from the earlier uint32 proto shape

+++ crates/coral-engine/src/contracts/query.rs
@@ -124,11 +124,11 @@ pub struct SourceValidationOutcome {
     /// One result per declared validation query, in manifest order.
     pub query_tests: Vec<QueryTestResult>,
     /// Total number of declared validation queries.
-    pub declared_query_count: u32,
+    pub declared_query_count: u64,
     /// Number of validation queries that executed successfully.
-    pub passed_query_count: u32,
+    pub passed_query_count: u64,
     /// Number of validation queries that failed.
-    pub failed_query_count: u32,
+    pub failed_query_count: u64,
     /// Whether every declared validation query passed.
     pub all_query_tests_passed: bool,
 }
@@ -137,9 +137,9 @@ impl SourceValidationOutcome {
     #[must_use]
     /// Builds one structured source-validation outcome.
     pub fn new(tables: Vec<super::TableInfo>, query_tests: Vec<QueryTestResult>) -> Self {
-        let declared_query_count = u32::try_from(query_tests.len()).unwrap_or(u32::MAX);
+        let declared_query_count = u64::try_from(query_tests.len()).unwrap_or(u64::MAX);
         let passed_query_count =
-            u32::try_from(query_tests.iter().filter(|test| test.passed).count()).unwrap_or(0);
+            u64::try_from(query_tests.iter().filter(|test| test.passed).count()).unwrap_or(0);
         let failed_query_count = declared_query_count.saturating_sub(passed_query_count);
         Self {
             tables,
